+++ src/common/layout.pug
@@ -1,0 +1,20 @@
+doctype html
+html
+  head
+    meta(charset='utf-8')
+    block title
+      title AO3 Enhancements
+    meta(name='viewport', content='width=device-width,initial-scale=1')
+    link(rel='icon', type='image/svg+xml', href='../icons/icon.svg')
+    link(rel='alternate icon', href='../icons/icon.ico')
+    link(
+      href='https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900',
+      rel='stylesheet'
+    )
+    block extrahead
+  body
+    if polyfill
+      script(
+        src='file-loader?name=[name].[ext]!./../../node_modules/webextension-polyfill/dist/browser-polyfill.min.js'
+      )
+    block scripts
