+++ src/options_ui/options_ui.pug
@@ -1,20 +1,11 @@
-doctype html
-html
-  head
-    meta(charset='utf-8')
-    title AO3 Enhancements Options
-    meta(name='viewport', content='width=device-width,initial-scale=1')
-    link(rel='icon', type='image/svg+xml', href='../icons/icon.svg')
-    link(rel='alternate icon', href='../icons/icon.ico')
-    link(
-      href='https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900',
-      rel='stylesheet'
-    )
-    if prod
-      link(href!='./options_ui.ts?entry&css', rel='stylesheet')
-  body
-    if polyfill
-      script(
-        src='file-loader?name=[name].[ext]!./../../node_modules/webextension-polyfill/dist/browser-polyfill.min.js'
-      )
-    script(src='./options_ui.ts?entry')
+extends ../common/layout.pug
+
+block title
+  title AO3 Enhancements Options
+
+block extrahead
+  if prod
+    link(href!='./options_ui.ts?entry&css', rel='stylesheet')
+
+block scripts
+  script(src='./options_ui.ts?entry')
