+++ src/reading_list/reading_list.pug
@@ -1,20 +1,11 @@
-doctype html
-html
-  head
-    meta(charset='utf-8')
-    title AO3 Enhancements Reading List
-    meta(name='viewport', content='width=device-width,initial-scale=1')
-    link(rel='icon', type='image/svg+xml', href='../icons/icon.svg')
-    link(rel='alternate icon', href='../icons/icon.ico')
-    link(
-      href='https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900',
-      rel='stylesheet'
-    )
-    if prod
-      link(href!='./reading_list.ts?entry&css', rel='stylesheet')
-  body
-    if polyfill
-      script(
-        src='file-loader?name=[name].[ext]!./../../node_modules/webextension-polyfill/dist/browser-polyfill.min.js'
-      )
-    script(src='./reading_list.ts?entry')
+extends ../common/layout.pug
+
+block title
+  title AO3 Enhancements Reading List
+
+block extrahead
+  if prod
+    link(href!='./reading_list.ts?entry&css', rel='stylesheet')
+
+block scripts
+  script(src='./reading_list.ts?entry')
