+++ types/notification-plugin.d.ts
@@ -1,4 +1,4 @@
-import { NotificationPlugin } from '@/options_ui/NotificationPlugin';
+import { NotificationPlugin } from '@/common/NotificationPlugin';
 
 declare module 'vue/types/vue' {
   export interface Vue {
